requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.9
Django==5.2.1
django-cors-headers==4.7.0  # ¡Esta línea sin comentarios al final!
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.3.0
filelock==3.18.0
gunicorn==23.0.0
kombu==5.5.3
packaging==25.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==6.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13